third_party/WebKit/LayoutTests/animations/api-readonly-object-types-expected.txt

This is a testharness.js-based test.
PASS KeyframeEffect constructor creates and exposes an AnimationEffectTiming object 
FAIL KeyframeEffectReadOnly constructor creates and exposes an AnimationEffectTimingReadOnly object assert_equals: expected "[object AnimationEffectTimingReadOnly]" but got "[object AnimationEffectTiming]"
Harness: the test ran to completion.
